unit tests for parseUri
